aaand another one: daysValid switch to Long

+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/utils/DatesHandler.kt
@@ -5,7 +5,7 @@ import java.util.*
 
 class DatesHandler {
 
-    private fun addDaysToDate(date: Date, days: Int): Date {
+    private fun addDaysToDate(date: Date, days: Long): Date {
         val calendar = Calendar.getInstance()
         calendar.time = date
         calendar.add(Calendar.DATE, days)
